requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
dj-inmemorystorage==2.1.0
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
future==0.18.2
idna==3.2
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
